+++ packages/date-picker/src/interface.ts
@@ -47,7 +47,15 @@ import type { DatePickerProps } from './props';
 
 export type DatePickerPlacementType = ComputedPlacement;
 
-export type DatePickerTypeType = 'year' | 'month' | 'monthrange' | 'date' | 'daterange' | 'datetime' | 'datetimerange';
+export type DatePickerTypeType =
+  | 'year'
+  | 'yearrange'
+  | 'month'
+  | 'monthrange'
+  | 'date'
+  | 'daterange'
+  | 'datetime'
+  | 'datetimerange';
 export type TimePickerTypeType = 'time' | 'timerange';
 export type PickerTypeType = DatePickerTypeType | TimePickerTypeType;