+++ packages/date-picker/src/utils.ts
@@ -146,6 +146,10 @@ export const typeValueResolver = {
     formatter: (value, format) => dateFormat(value, format),
     parser: (text, format) => fecha.parse(text, format || 'yyyy-MM-dd'),
   },
+  yearrange: {
+    formatter: rangeFormatter,
+    parser: rangeParser,
+  },
   multiple: {
     formatter(value, format) {
       return value
@@ -249,6 +253,7 @@ export const DEFAULT_FORMATS: Record<PickerTypeType, string> = {
   month: 'yyyy-MM',
   monthrange: 'yyyy-MM',
   year: 'yyyy',
+  yearrange: 'yyyy',
   datetime: 'yyyy-MM-dd HH:mm:ss',
   time: 'HH:mm:ss',
   timerange: 'HH:mm:ss',
@@ -419,7 +424,15 @@ export const isInRange = (time, a, b) => {
   if (!a || !b) {
     return false;
   }
-  const [start, end] = [a, b].sort();
+  const [start, end] = [a, b].sort((x, y) => {
+    if (x - y > 0) {
+      return 1;
+    }
+    if (x - y < 0) {
+      return -1;
+    }
+    return 0;
+  });
   return time >= start && time <= end;
 };